+++ packages/hono-takibi/src/generators/zod/schema/generate-zod-schema.ts
@@ -6,7 +6,7 @@ import { isFormatString } from '../../../core/validator/is-format-string'
 import { generateZodNumberSchema } from '../generate-zod-number-schema'
 import { generateZodIntegerSchema } from '../generate-zod-integer-schema'
 import { generateAllOfCode } from '../../hono/openapi/component/allof/generate-allof-code'
-import { generateAnyOfCode } from '../../openapi/components/anyof/generate-anyof-code'
+import { generateAnyOfCode } from '../../hono/openapi/component/anyof/generate-anyof-code'
 import { generateOneOfCode } from '../../openapi/components/oneof/generate-oneof-code'
 import { getVariableSchemaNameHelper } from '../../../core/helper/get-variable-schema-name-helper'
 import { generateZodObject } from '../generate-zod-object'
